+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## UNRELEASED
 
+### Changes
+
+* Rich text styles are now split into Block and Inline controls for a better UX when applying text styles.
+
 ### Adds
 
+* Add support where classes defined for rich text styles can be toggled on multiple RT elements.
 * Add translation keys used by the multisite assembly module.
 * Add side by side comparison support in AposSchema component.
 * Add the possibility to make widget modals wider, which can be useful for widgets that contain areas taking significant space. See [documentation](https://v3.docs.apostrophecms.org/reference/modules/widget-type.html#options).