+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
-## UNRELEASED
+## 4.0.0 (2024-03-12)
 
 ### Adds
 
 * Add translation keys used by the multisite assembly module.
 * Add side by side comparison support in AposSchema component.
 * Add the possibility to make widget modals wider, which can be useful for widgets that contain areas taking significant space. See [documentation](https://v3.docs.apostrophecms.org/reference/modules/widget-type.html#options).
+
+### Changes
+
+* Migrate to Vue 3.
 
 ### Fixes